more api stuff

+++ app/Http/Controllers/CategoryAPIController.php
@@ -18,34 +18,44 @@ class CategoryAPIController extends Controller
     public function store(Request $request)
     {
 		$validator = Validator::make($request->all(), (new storeCategory)->rules());
-		if ($validator->fails()) 
+		if ($validator->fails())
 		{
 			return response()->json($validator->errors(), 400);
-		} else 
+		} else
 		{
 			$category = Categories::create($request->all());
 			return response()->json($category, 201);
-		}			
+		}
     }
 
 	public function update(Request $request)
     {
 		$validator = Validator::make($request->all(), (new storeCategory)->rules());
-		if ($validator->fails()) 
+		if ($validator->fails())
 		{
 			return response()->json($validator->errors(), 400);
-		} else 
+		} else
 		{
-			$category = Categories::find($request['id']);
+			$category = Categories::find($request['category_id']);
+
+            if (is_null($category))
+            {
+                return response()->json(null, 404);
+            }
+
 			$category->update($request->all());
 			return response()->json($category, 200);
-		}	
-		
+		}
+
     }
 
     public function destroy(Request $request)
     {
-        $category = Categories::find($request['id']);
+        $category = Categories::find($request['category_id']);
+        if (is_null($category))
+        {
+            return response()->json(null, 404);
+        }
 		$category->delete();
 		return response()->json(null, 204);
     }
