+++ app/Http/Controllers/CountryAPIController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Countries;
+use App\Country;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\DB\View;
 use Cabon\Carbon;
@@ -18,25 +18,31 @@ class CountryAPIController extends Controller
     public function store(Request $request)
     {
 		$validator = Validator::make($request->all(), (new storeCountry)->rules());
-		if ($validator->fails()) 
+		if ($validator->fails())
 		{
 			return response()->json($validator->errors(), 400);
-		} else 
+		} else
 		{
-			$country = Countries::create($request->all());
+			$country = Country::create($request->all());
 			return response()->json($country, 201);
-		}	
+		}
     }
 
 	public function update(Request $request)
     {
 		$validator = Validator::make($request->all(), (new storeCountry)->rules());
-		if ($validator->fails()) 
+		if ($validator->fails())
 		{
 			return response()->json($validator->errors(), 400);
-		} else 
+		} else
 		{
-			$country = Countries::find($request['id']);
+			$country = Country::find($request['country_id']);
+
+            if (is_null($country))
+            {
+                return response()->json(null, 404);
+            }
+
 			$country->update($request->all());
 			return response()->json($country, 200);
 		}
@@ -45,8 +51,14 @@ class CountryAPIController extends Controller
 
     public function destroy(Request $request)
     {
-        $country = Countries::find($request['id']);
+        $country = Country::find($request['country_id']);
+
+        if (is_null($country))
+        {
+            return response()->json(null, 404);
+        }
+        
 		$country->delete();
-		return response()->json(null, 204);
+		return response()->json(null, 200);
     }
 }
