+++ app/Http/Controllers/RestaurantAPIController.php
@@ -3,9 +3,12 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Restaurants;
+use App\Http\Requests\RestaurantRequest;
+use App\Restaurant;
 use App\Posts;
 use App\Comments;
+use App\Categories;
+use App\Country;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\DB\View;
 use Cabon\Carbon;
@@ -22,26 +25,30 @@ class RestaurantAPIController extends Controller
 {
     public function store(Request $request)
     {
-		$validator = Validator::make($request->all(), (new storeRestaurant)->rules());
-		if ($validator->fails()) 
+		$validator = Validator::make($request->all(), (new RestaurantRequest)->rules());
+		if ($validator->fails())
 		{
 			return response()->json($validator->errors(), 400);
-		} else 
+		} else
 		{
-			$restaurant = Restaurants::create($request->all());
+			$restaurant = Restaurant::create($request->all());
 			return response()->json($restaurant, 201);
 		}
     }
 
 	public function update(Request $request)
     {
-		$validator = Validator::make($request->all(), (new storeRestaurant)->rules());
-		if ($validator->fails()) 
+		$validator = Validator::make($request->all(), (new RestaurantRequest)->rules());
+		if ($validator->fails())
 		{
 			return response()->json($validator->errors(), 400);
-		} else 
+		} else
 		{
-			$restaurant = Restaurants::find($request['restid']);
+			$restaurant = Restaurant::find($request['restaurant_id']);
+            if (is_null($restaurant))
+            {
+                return response()->json(null, 404);
+            }
 			$restaurant->update($request->all());
 			return response()->json($restaurant, 200);
 		}
@@ -50,46 +57,50 @@ class RestaurantAPIController extends Controller
     public function destroy(Request $request)
     {
 		$validator = Validator::make($request->all(), (new deleteRestaurant)->rules());
-		if ($validator->fails()) 
+		if ($validator->fails())
 		{
 			return response()->json($validator->errors(), 400);
-		} else 
+		} else
 		{
-			$restaurant = Restaurants::find($request['restid']);
+			$restaurant = Restaurant::find($request['restaurant_id']);
+            if (is_null($restaurant))
+            {
+                return response()->json(null, 404);
+            }
 			$restaurant->delete();
 			return response()->json(null, 204);
-		}	
+		}
     }
-	
-	public function showposts(Request $request)
+
+	public function showposts($id)
     {
-		$validator = Validator::make($request->all(), (new deleteRestaurant)->rules());
-		if ($validator->fails()) 
-		{
-			return response()->json($validator->errors(), 400);
-		} else 
-		{
-			$restaurants = Restaurants::find($request['restid']);
-			$restaurants->posts = $restaurants->posts->all();
-			foreach($restaurants->posts as $post)
-			{
-					$post->comments = $post->comments->all();
-			}
-			return response()->json($restaurants, 200);
-		}	
+        $restaurants = Restaurant::find($id);
+
+        if (is_null($restaurants))
+        {
+            return response()->json($id, 404);
+        }
+
+        $restaurants->posts = $restaurants->posts->all();
+
+        foreach($restaurants->posts as $post)
+        {
+            $post->comments = $post->comments->all();
+        }
+
+        return response()->json($restaurants, 200);
+
     }
-	
-    public function showlocationcategory(Request $request)
+
+    public function showcategory($id)
     {
-		$validator = Validator::make($request->all(), (new showRestaurant)->rules());
-		if ($validator->fails()) 
-		{
-			return response()->json($validator->errors(), 400);
-		} else 
-		{
-			$restaurants = Restaurants::where('category_id', $request['category_id'])->where('country_id', $request['country_id'])->get();
-			return response()->json($restaurants, 200);
-		}	
-		
+        $restaurants = Categories::find($id)->restaurants;
+		return response()->json($restaurants, 200);
+    }
+
+    public function showCountry($id)
+    {
+        $restaurants = Country::find($id)->restaurants;
+		return response()->json($restaurants, 200);
     }
 }
